separate files and cleanup code
