+++ CMakeLists.txt
@@ -25,8 +25,8 @@ list(APPEND CUDA_NVCC_FLAGS
 
 CUDA_ADD_LIBRARY(host_to_device host_to_device.cu)
 
-add_executable(main main.cpp) 
-target_include_directories(main PRIVATE ${CUDA_TOOLKIT_INCLUDE})
+add_executable(main main.cpp allocation.cpp) 
+target_include_directories(main PRIVATE ${CUDA_TOOLKIT_INCLUDE} allocation.hpp)
 
 target_link_libraries(main PRIVATE host_to_device)
