+++ allocation.hpp
@@ -1,0 +1,6 @@
+
+int** alloc_2d_init(int r, int c);
+
+void data_init(int** A, int r, int c);
+
+void print_helper(int** A, int r, int c);
